SDK-1585: Added checking existence of configuration index Fixed tests_files

+++ src/ManifestStrategy/ConfigureEnvManifestStrategy.php
@@ -77,7 +77,9 @@ class ConfigureEnvManifestStrategy extends AbstractManifestStrategy
             return false;
         }
         if (!$isDry) {
-            file_put_contents($configFileName, $this->getConfigAppendData($target, $value), FILE_APPEND);
+            if (!$this->isTargetExist($target, $inputOutput)) {
+                file_put_contents($configFileName, $this->getConfigAppendData($target, $value), FILE_APPEND);
+            }
         }
 
         $inputOutput->writeln(sprintf(
@@ -88,6 +90,29 @@ class ConfigureEnvManifestStrategy extends AbstractManifestStrategy
         ), InputOutputInterface::DEBUG);
 
         return true;
+    }
+
+    /**
+     * @param string $target
+     * @param \SprykerSdk\Integrator\Dependency\Console\InputOutputInterface $inputOutput
+     *
+     * @return bool
+     */
+    protected function isTargetExist(string $target, InputOutputInterface $inputOutput): bool
+    {
+        $configFileName = $this->config->getConfigPath();
+        $configFileContent = file_get_contents($configFileName);
+
+        if ($configFileContent === false) {
+            $inputOutput->writeln(sprintf(
+                'Could not read from file `%s`. Please check file path or customize using `IntegratorConfig::getConfigPath()`.',
+                $configFileName,
+            ), InputOutputInterface::DEBUG);
+
+            return false;
+        }
+
+        return str_contains($configFileContent, $this->getConfigTarget($target));
     }
 
     /**
@@ -104,6 +129,16 @@ class ConfigureEnvManifestStrategy extends AbstractManifestStrategy
         $data .= ';' . PHP_EOL;
 
         return $data;
+    }
+
+    /**
+     * @param string $target
+     *
+     * @return string
+     */
+    protected function getConfigTarget(string $target): string
+    {
+        return '$' . $this->config->getConfigVariableName() . '[' . $target . ']';
     }
 
     /**
